refactor(solver): rename must_satisfy → parent_paramount
The "must satisfy ONE of anything" semantics is being replaced by hard
MP-only enforcement in #1379. Renaming the constraint module and its
tests to match the new policy name (used by the wife-feedback scoreboard
and prior parent-paramount stages).

Behavior unchanged in this commit — pure rename + import updates.
The diagnostic log keys (must_satisfy_one_no_possible / _material_parent_unmet
/ _other_unmet) stay as-is to preserve the dashboard contract; only the
module/file/test paths move.

The test_must_satisfy_one_cleanup.py file kept its filename (it pins
config-cleanup history, not the constraint module).

Refs #1379.
