+++ tests/unit/solver/test_parent_paramount_diagnostic.py
